auth working on the CRUD

+++ controllers/stuff.js
@@ -1,33 +1,83 @@
 const Thing = require("../models/Thing");
+const fs = require("fs");
 
 exports.createThing = (req, res, next) => {
-  delete req.body._id;
+  const thingObject = JSON.parse(req.body.thing);
+  delete thingObject._id;
+  delete thingObject._userId;
   const thing = new Thing({
-    // title: req.body.title,
-    // il y a un raccourci ci-dessous.
-    ...req.body
+    ...thingObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
   });
+
   thing
     .save()
-    .then(() => res.status(201).json({ messega: "Objet enregistré" }))
-    .catch(error => res.status(400).json({ error }));
+    .then(() => {
+      res.status(201).json({ thing });
+    })
+    .catch(error => {
+      res.status(400).json({ error });
+    });
 };
 exports.getOneThing = (req, res, next) => {
   Thing.findOne({ _id: req.params.id })
     .then(thing => res.status(200).json(thing))
+    // la ligne ci-dessus renvoie l'objet crée.
     .catch(error => res.status(404).json({ error }));
 };
 
 exports.modifyThing = (req, res, next) => {
-  Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json("Objet modifié"))
-    .catch(error => res.status(404).json({ error }));
+  const thingObject = req.file
+    ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`
+      }
+    : { ...req.body };
+
+  delete thingObject._userId;
+  Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+      if (thing.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non authorisé" });
+      } else {
+        Thing.updateOne(
+          { _id: req.params.id },
+          { ...thingObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ thingObject }))
+          // la ligne ci-dessus renvoie l'objet modifié.
+          .catch(error => res.status(401).json({ error }));
+      }
+    })
+    .catch(error => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json("Objet supprimé"))
-    .catch(error => res.status(404).json({ error }));
+  Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+      if (thing.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        const filename = thing.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Thing.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            // .then(() => { res.status(200).json({thing})})
+            // cette ligne doit être testée.
+            .catch(error => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch(error => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getAllThings = (req, res, next) => {
